Basic packet counter
Every packet that gets "processed" is allocated a the next number
available so it can be tracked through the logs. This counts the total
packets received, before any validation or decoding, the packet id
counter resets to zero when emonhub restarted. Currently the id is left
attached to end of each data string, will probably be removed before
release.