Configuration file section checks
At start up and at every change to the settings file it is checked for
the 3 compulsory sections. At start up up emonhub cannot start but will
point user to the section causing an issue/missing. During normal
running if a user save the file with one of the sections missing the
changes are ignored until rectified.