Serial port and socket defaults
These defaults are more about allowing emonHub to continue to start when
serial port or socket port numbers have been omitted or been incorrectly
set or edited.  so rather than just the blank place holder I included
the Pi serial port address as the JeeListener default and the 'usual
default' of 50011 for the socket port number.