Pause 'in' discards frame
The pause 'in' now works slightly differently on the serial and socket
inputs to avoid any unprocessed out of date or inaccurate frames being
read from the inputs buffer after a pause is lifted. The pause 'in' now
allows the input to function normally but discards the frame and clears
the buffer before any processing occurs.