Setting timestamp optional (corrected comment)
Setting timestamp optional  Removed the need to have the timestamp code
in each read() method by automatically adding it if it is omitted at the
start of _process_frame(). less repetitive code and less mandatory items
for custom interfacers to have. I have left JeeInterfacer with the
timestamp code in read() as an example but it is not mandatory as shown
in the SocketInterfacer read().