Amendment to serial response test
Changed response test to check for "\x00" in addition to "\r\n" as it
seems the first response from an rfm2pi from a cold start fails the
original "contains \r\n" test due to a stray "\x00", this results in the
serial port not being opened. As this is only apparent in 1st response
the original test is retained for subsequent restarts and checks.