Jee device default settings
OEM project defaults of baseid = 15, group = 210 and frequency = 433 are
set by default. quiet is also set to True by default. If a settings
string is returned from Jee device at set up it updates "info" and this
string is used to check if sending a command setting is actually
required, each time a command is send the settings status "info" is
updated. Settings are sent to device if status is unconfirmed so that
devices with earlier firmwares get set regardless at start up, as was
previously done with all devices. All default settings are now strings
to resolve variable type issues where a native type gets cast to a
string when a value is passed from settings file.