Properly rewrite VEDirect interfacer for Python3.
The main problem was trying to decode the checksum byte (0-255) as utf-8
(0-127 plus continuation bytes), so we have to store everything as bytes
until we are ready to decode it.

I also took the opportunity to rewrite a few things in a simpler / more
pythonic style. This might raise different exceptions when things go
wrong, but it is shorter and it should be simpler to understand. I
tried to add some helpful comments!