Set timeouts on requests transactions.
Requests will wait forever for a response once the server has opened the
connection. This might cause hard to detect bugs. Set a ridiculously
long timeout. From the docs:

timeout is not a time limit on the entire response download; rather, an
exception is raised if the server has not issued a response for timeout
seconds (more precisely, if no bytes have been received on the
underlying socket for timeout seconds). If no timeout is specified
explicitly, requests do not time out.