Create emonhub.service.bullseye
Create a service file for use with legacy Bullseye on PiZero without the requirement for `var-log.mount`. Do not understand why it is not made available, but it isn't.

Cannot be done via an add-in as you cannot remove requirements, just add them.

Could be automated via the install by checking hardware (as we do elsewhere) but low use case.